API

SBMLGraphs.SBMLGraphs — Module

SBMLGraphs.jl

Package for converting Systems Biology Markup Language (SBML) XML files into common graph representations.

Extensions

SparseArrays.jl

Graphs.jl

Convertion from SBML to other types

Base.convert — Method

Base.convert(t::Type{Dict}, species::SBML.Species)

Convert an SBML.Species object to a Dictionary.

Arguments

t::Type{Dict}: The target type for conversion (Dictionary).

reaction::SBML.Species: The SBML.Reaction object to be converted.

Returns

Dict: A Dictionary containing the fields of the SBML.Reaction object as key-value pairs.

Base.convert — Method

Base.convert(t::Type{Dict}, s::SBML.Reaction)

Convert an SBML.Reaction object to a Dictionary.

Arguments

t::Type{Dict}: The target type for conversion (Dictionary).

s::SBML.Reaction: The SBML.Reaction object to be converted.

Returns

Dict: A Dictionary containing the fields of the SBML.Reaction object as key-value pairs.

Base.convert — Method

Base.convert(_::Type{AbstractMatrix}, m::SBML.Model)

Arguments

t::Type{AbstractMatrix}: The target type for conversion (AbstractMatrix).

s::SBML.Model: The SBML.Reaction object to be converted.

Returns

M::AbstractMatrix{Bool}: Adjacency matrix for the directed bipartite reaction-species graph contructed from the model.

V::AbstractVector{String}: Node identifiers.

Relevant graph projections

SBMLGraphs.projected_graph — Method

projected_graph(A::AbstractMatrix{T}, Vp::AbstractVector{Int}) where T

Returns the projection of A onto one of its node sets.

Returns the adjacency matrix Ap that is the projection of the bipartite graph adjacency matrix A onto the specified nodes. They are connected in Ap if they have a common neighbor in A.

Arguments

A::AbstractMatrix{T}: Adjacency matrix of the bipartite graph.

Vp::AbstractVector{Int}: List of nodes to project onto.

Returns

AbstractMatrix{T}: A graph adjacency matrix that is the projection onto the given nodes.

Extended help

This algorithm doesn't check if the inputted adjacency matrix is for a bipartite graph, therefore the algorithm will fail whenever it detects it's not bipartite, resulting in useless computation time.

References

projected_graph function from networkx

SBMLGraphs.get_species_graph — Method

get_species_graph(model::SBML.Model, A::AbstractMatrix, V::AbstractVector{String})

Contructs the species-centered graph based on SBML model, adjacency matrix and node identifiers.

Arguments

m::SBML.Model: The SBML model to convert.

A::AbstractMatrix{T}: Adjacency matrix of graph.

Vp::AbstractVector{Int}: List of nodes identifiers.

Returns

AbstractMatrix{T}: Projected graph adjacency matrix.

AbstractVector{String}: Species nodes identifiers.

SBMLGraphs.get_reactions_graph — Method

get_reactions_graph(model::SBML.Model, A::AbstractMatrix, V::AbstractVector{String})

Contructs the reaction-centered graph based on SBML model, adjacency matrix and node identifiers.

Arguments

m::SBML.Model: The SBML model to convert.

A::AbstractMatrix{T}: Adjacency matrix of graph.

Vp::AbstractVector{Int}: List of nodes identifiers.

Returns

AbstractMatrix{T}: Projected graph adjacency matrix.

AbstractVector{String}: Reaction nodes identifiers.