Graphs.jl

Base.convert — Method

Base.convert(t::Type{Graphs.Graph}, m::SBML.Model)::Tuple{Graphs.Graph,AbstractVector{String}}

Convert an SBML model instance to a Graph instance and the nodes identifiers.

Arguments

t::Type{Graphs.Graph}: The type of Graph to convert to.

m::SBML.Model: The SBML model to be converted.

Returns

Graphs.Graph: SBML model undirected graph instance.

AbstractVector{String}: Node identifiers.

Base.convert — Method

Base.convert(t::Type{Graphs.DiGraph}, m::SBML.Model)::Tuple{Graphs.DiGraph,AbstractVector{String}}

Convert an SBML model instance to a DiGraph instance and the nodes identifiers.

Arguments

t::Type{Graphs.DiGraph}: The type of Graph to convert to.

m::SBML.Model: The SBML model to be converted.

Returns

Graphs.DiGraph: SBML model directed graph instance.

AbstractVector{String}: Node identifiers.

SBMLGraphs.projected_graph — Method

SBMLGraphs.projected_graph(A::Graphs.AbstractGraph, Vp::AbstractVector{Int}) where T

Returns the projection of G onto one of its node set.

Arguments

G::Graphs.AbstractGraph: Bipartite graph.

Vp::AbstractVector{Int}: List of nodes to project onto.

Returns

Graphs.AbstractGraph: A graph that is the projection onto the given nodes.

SBMLGraphs.get_species_graph — Method

SBMLGraphs.get_species_graph(model::SBML.Model, G::Graphs.AbstractGraph, V::AbstractVector{String})

Contructs the species-centered graph based on SBML model, graph and node identifiers.

Arguments

m::SBML.Model: The SBML model to convert.

A::AbstractGraph: Graph.

Vp::AbstractVector{Int}: List of nodes identifiers.

Returns

AbstractGraph: Projected graph over species nodes.

AbstractVector{String}: Species nodes identifiers.

SBMLGraphs.get_reactions_graph — Method

SBMLGraphs.get_reactions_graph(model::SBML.Model, G::Graphs.AbstractGraph, V::AbstractVector{String})

Contructs the reactions-centered graph based on SBML model, graph and node identifiers.

Arguments

m::SBML.Model: The SBML model to convert.

A::AbstractGraph: Graph.

Vp::AbstractVector{Int}: List of nodes identifiers.

Returns

AbstractGraph: Projected graph over reaction nodes.

AbstractVector{String}: Reaction nodes identifiers.

SparseArrays.jl